Styles file for profile screen added
