+++ components/CloseIcon.js
@@ -1,11 +1,12 @@
 import React from "react";
 import { TouchableOpacity } from "react-native";
 import AntDesign from "react-native-vector-icons/AntDesign";
+import { colorsTheme } from "../styles/colorsTheme";
 
 const CloseIcon = ({ onPress }) => {
   return (
     <TouchableOpacity onPress={onPress}>
-      <AntDesign name="close" size={22} color="black" />
+      <AntDesign name="close" size={22} color={colorsTheme.darkBlue} />
     </TouchableOpacity>
   );
 };
